Remove the OTEL_COLLECTOR_HOST variable (#89)
Remove the usage of, and logic surrounding the OTEL_COLLECTOR_HOST
variable. As a way to to uniform how we reference the OpenTelemetry
Collector, we are sending the OTEL_EXPORTER_OTLP_ENDPOINT directly to
the ESR job and thus removing the need of using OTEL_COLLECTOR_HOST as
an intermediary guard rail variable.